requirements.txt
numpy>=1.26
opencv-python>=4.7
Pillow>=10.0
PyYAML>=6.0
timm>=1.0
torch>=2.2
torchvision>=0.17
tqdm>=4.66
